Disable automatic `return None` for non-Nonetype.
We don't need it since mypy won't let us write code that doesn't
return in the first place (if the type is not None). But this creates
another problem - we will end up with a basic block with no branching
at the end.

To get around this, add 'Unreachable' whose only job is to signal the
block formatter that its okay for there to not be a leave in the block.
It is completely ignored by all visitors except analysis, for which it
acts (most of the time) like a Return.