Add new types Register, Label, AnalysisDict, AnalysisResult
Of these only AnalysisResult is anything more complicated than a newtype
(due to a bug in libtyping which exists on older versions of python still
in use). Register and Label are the focus of this patch and are newtypes.

This aids in code readability and correctness immensely as its possible
to tell by the type signature whether something is a label or a register
(since before they were both ints).