Add support for dict update
Originally I thought this would be a simple change to get
reacustomed to the codebase. Turns out I had to go down a
rabbithole because is_void binary ops are broken, but
binary operations that mutate the first argument are common
enough that I think its worth fixing this on a general level.

As a side note, I saw an instance where we pass INVALID_REGISTER
as an argument to a function which takes Optional[Register].
I *think* this is a bug, because in refcount.py we only avoid
dealing with the dest register if its None; we don't have a
INVALID_REGISTER check. It errored when I passed in INVALID_REGISTER
(because I was cargo culting). I'm not sure why the other usage
isn't failing but I can send in another PR to fix that to None
if it doesn't break anything...