Add support for dict update (#54)
Originally I thought this would be a simple change to get
reacustomed to the codebase. Turns out I had to go down a
rabbithole because is_void binary ops are broken, but
binary operations that mutate the first argument are common
enough that I think its worth fixing this on a general level.