Add a travis build that uses a debug build of python
Debug builds of python are more likely to crash or assert in response
to memory errors such as refcounting mistakes, which will make it easier
to catch errors.

Since the debug build of python is slower, we only run tests involving
C code that we are responsible for: unit tests of our runtime and the
`run` tests.

Unfortunately the process for getting a debug build of python into our
travis environment is unbelievably janky:
 * Build Python 3.6.5 with: `./configure CFLAGS='-DPy_DEBUG -DPy_TRACE_REFS -DPYMALLOC_DEBUG' --with-pydebug --prefix=/home/travis/python-3.6.5-debug`
 * Install that python and create a virtualenv in
   `/home/travis/virtualenv/python-3.6-debug/', which mimics what the
   official travis Pythons do and makes it easy use the right python
   version.
 * Tar up /home/travis and stick it up on a web server
   somewhere. Right now I am just personally hosting it because it was
   the most expedient thing, but there are certainly better places to
   host it.
 * In the travis build, wget the tarball and unpack it.