Add travis builds that use debug builds of python
Debug builds of python are more likely to crash or assert in response
to memory errors such as refcounting mistakes, which will make it easier
to catch errors.

Since the debug build of python is slower, we only run tests involving
C code that we are responsible for: unit tests of our runtime and the
`run` tests.

The build works by downloading a python source tarball and building it
on travis.  While this is not as slow as one might imagine (about 5m),
it is still too slow to be reasonable, so Travis's caching is used to
avoid needing to repeat the build.