Fix some bugs with misaimed targets
`alloc_target()` shouldn't be used from code that is not certain
to own the current target. This was causing trouble with code like
`x = a.f()`, where "f" was being assinged to `x`.

Fix that, and add some slight infrastructure for optionally specifying
targets to functions that may or may not own the target.