Refactor primitive ops (#88)
Reimplement primitive ops. The main idea is to make the ops data-driven,
so that new ops can be defined in a declarative fashion, and simple ops 
don't require any non-trivial code to be written. Additionally, ops can be 
defined in a modular fashion, with separate modules for ops related to 
each primitive type (e.g. `mypyc.ops_list`). For most primitive ops, no 
changes to `mypyc.genops` or `mypyc.emitfunc` are needed any more.

Also move `NEW_TUPLE` to a separate class, since it's not a good fit as
it's polymorphic, and primitive ops are supposed to be "monomorphic".

The changes to test cases should all be cosmetic.

Coincidentally this breaks int negation, since we no longer translate that
to `0 - x`. I'll fix it in a separate PR. Nothing seems to depend on it working
right now.

Fixes #59.