Always do basic tests for benchmark tests
Not running the benchmark tests at *all* in the normal
testing mode makes it too easy to break them.
Instead the benchmarks should be set up to be runnable in both
a benchmarking and non-benchmarking mode.

We signal the benchmarking mode with an environment variable.

This is all kind of hacky but should be fine for now.