Support For-Loops for Arbitrary Iterables
Implement the iter and next operators so that iterables can be
iterated using 'for'.

A prelimary that needed to be done was fix a
bug when generating IR for continue and break statements, where
duplicate `goto` instructions were being generated. To fix this bug,
a dictionaries storing the positions of the continue and break
statements in the blocks were created so that all control-flow ops
following a continue or break are removed from the list of
instructions.

This fixes #100.