Support For-Loops for Arbitrary Iterables
Implement the iter and next operators so that iterables can be
iterated using 'for'.

A prelimary that needed to be done was fix a bug when generating
IR for continue and break statements, where duplicate `goto`
instructions were being generated. To fix this bug, a new block is
now inserted immediately after the 'goto' instruction associated
with every continue and break statement.

This fixes #100.