Eliminate PyGetAttr/PySetAttr and turn them into primitive ops
It turns out that using Python_GetAttr/Python_SetAttr with pre-created
static strings is a lot faster than using Python_GetAttrString and
Python_SetAttrString (since those need to convert the C string into a
Python string). I made this change for GetAttr in #146 (but somehow
forgot about SetAttr??).

But now that PyGetAttr and PySetAttr were just simple operations on a
list of Values, there is no reason they need to be their own Ops
anymore, so we can turn them into PrimitiveOps and delete a bunch of
code.