Refactor Function Class Instantiation
A callable class representing a function is now instantiated when
the function is defined, not when it is called. This removes
duplicate instantiations of the same callable classes when really
only a single instantiation is needed.

In addition, the environment symbol table can now take FuncDef node
types in addition to Var node types.