Make trait method dispatch efficient (#201)
Emit side vtables for each trait that a class implements and store
them in an array placed before the regular vtable. To call the
methods, then, we (at runtime) search the array for the the trait
vtable and dispatch through it.
(This may not seem all that efficient, but the alternative is a Python
method dispatch, so.)

Changing the visitor to be a trait in the TreeVisitor benchmark shows
minimal perf degradation. (And a 5x perf degredation without this PR!)