Evaluate statements on module top levels
Create a hidden function that corresponds to the module top level.
Also add support for `global` declarations.

Module-level attributes are accessed through the globals dictionary,
which makes them pretty slow.

Imports are still treated as declarations and they are evaluated
before evaluating non-import statements at the module top level. We
can fix this later but the current approach is probably fine for a
while.