Added set literals (#267)
Set literals are created by calling `PySet_New`, then adding each element in the literal with `PySet_Add`. There isn't a varargs constructor for sets, and generating a different iterable, such as a list, and passing that to the constructor will probably do the same thing in the background.
Sets are still treated as generic objects by this change.