Added `__getitem__` support (#290)
Index expressions in native code on natively compiled classes now call the native `__getitem__` method instead of using `PyObject_GetItem`, removing a layer of indirection.

Native classes that have a `__getitem__` method now attach that method into the `tp_as_mapping` field of the Python type struct. This allows external objects to treat them as subscriptable, and also allows `PyObject_GetItem` to be used on them in the case of things like `Any`.
This means that external code can treat internal objects as subscriptable.

This adds code to `genops`, in `visit_index_expr`, to replace the `PyObject_GetItem` special method call with a native method call. This looks similar to the code in `translate_method_call`, but `translate_method_call` attemps a special method call before a native method call, where in this case we prefer to do the other order.

This also adds code to `emitclass`, `emitmodule`, and `emitwrapper`. The code in `emitclass` creates a `PyMappingMethods` struct and populates it with the `GetItem` method. The code in `emitmodule` emits wrappers for each method in a given list, which currently contains only `__getitem__`. The code in `emitwrapper` contains the code for actually producing that wrapper. This was very similar to code for the existing `PyPy` wrapper, so this was refactored a bit so that both methods can call the same core.

Closes #178.