Introduce a `FuncDecl` abstraction and use it (#316)
Add a notion of function declarations, represented by a `FuncDecl` object.
`FuncDecl` contains the signature, name, class, etc. Basically everything
from a `FuncIR` except the environment and code.

We prepopulate tables of function declarations for both functions and methods before
we start generating code, and we change native function and method calls to be driven
by the types from the `FuncDecl` instead of the types from mypy.
The `Call` instruction now takes a `FuncDecl` to represent the function to call instead
of a name.

The proximate reason for this refactor is so that I could implement `super()`
without feeling like I needed to take a shower.