Allocate types on the heap
CPython overloads `Py_TPFLAGS_HEAPTYPE` to mean essentially two different things:
1. The type object is literally allocated on the heap
   (and is a PyHeapTypeObject)
2. The type is not a "builtin" type and thus is allowed to support a number
   of operations that builtin types can not, such as allowing attributes of
   the type to be set.

We don't particularly care about #1, but do care about #2, so here we are.

The "correct" interface to use for creating heap types is
PyType_FromSpec(), but that has a *completely* different API and
doesn't exist at all in 2.7, which we still may want to support some
day.
Instead we manually create the type and fill in some fields ourselves.