Rework the handling of type objects to use pointers directly (#318)
This is to lay the groundwork for allocating our TypeObject's on the
heap, which we will unfortunately need to do to support a bunch of
behavior.

The main tricky thing here is in the initialization of trait vtables,
which can't be done fully statically. We take the approach of filling
out the initial vtable with pointers to the variable where the pointer
will be stored, and then patching them all up at module load
time. This is a bummer.

A side effect of these changes is that we no longer allow two modules
to both define classes that inherit from a class defined in the
other. This is fine, cpython doesn't support it either.