Support throw method for generator functions (#370)
To do this, a helper function is generated that the __next__ and throw methods can call. The helper function takes in type, value, and traceback arguments corresponding to a possible exception thrown, and checks if these arguments are Py_None every time it is called. When __next__ calls the helper function, it passes in Py_None, so no exception is raised. When throw calls the helper function, it passes in the arguments passed in by the user, so that an exception is first thrown, and then the program executes until the next yield statement or raises StopIteration if the exception is not caught.

This fixes #345.