Overhaul the mechanism for building C (#435)
This largely imitates the Cython approach, AIUI.

Instead of having `mypyc.buildc` generate a bunch of `setup.py` files, we provide a
`mypycify` routine that can be invoked from *within* a distutils or setuptools `setup.py`
in order to generate the C code and produce `Extension` objects.

In order to be able to run `install_name_tool` after the shared library is built on OS X,
we need to subclass the `build_ext` command so that we can hook it after a build is done.
Since we're doing that anyways we also take advantage of this to easily do a few things
that we probably could have manage in another way (figuring out the file names of libraries
being generated in advance).

The recommended way, then, for projects to interact with mypyc, is to call
`mypycify` from a `setup.py`, since this will support building, installing, and
creating packages.

`mypycify` then becomes a simple wrapper script that generates a `setup.py`
that calls `mypy.build.mypycify`. 

Fixes #431.