Fix problems in the test suite that caused compile errors to be silent
Unfortunately `run_setup` ignores errors. We could go back to using
subprocess but that slows down the tests by 30s on my machine, which
is real bad. (It is worse now that it used to be, I think, since
the setup.py needs to import most of mypy and mypyc.)