Implement multi-file compilation (#510)
This implements an initial version of multi-file compilation. Each module is
turned into its own .c file. A shared `__native.c` contains the globals initialization
and global definitions.

The major changes needed to support this are:
1. Declarations used by multiple files are now all collected into a single
    header file (which is 27kloc for mypy). 
2. Most of the order dependence in the code needed to be removed,
    so that different modules could be untangled.
3. This meant adding more forward declarations of things. `HeaderDeclaration`
    was generalized to contain both a declaration (that goes into the shared header)
    and an optional definition (that goes into a shared `.c` file)
4. The runtime finally needs its own `.c` file to prevent duplication of data structures
    it uses. As a TODO, more stuff should be moved into it from the header.

Compilation on Linux with clang is made
slightly *slower* (though that could change if we made it parallelizable).
On Windows, it seems to help some but not enormously: it is not enough to 
get the build to not OOM on a free Appveyor, but seems like it might keep
the build from intermittently OOMing on a "quad VM" Appveyor.

Initial numbers showed a slowdown (like 5%), but I haven't dug into it too deeply yet.

I am inclined to merge this PR anyways, though, even if we don't turn on multi-file by
default. The abstractions for generating multiple files are probably good, and I think
most of the changes needed to make it work are also good.

We run multi-module tests in both modes.

The default remains off.