Declare all the globals in header files as "extern"
Apparently on macOS it doesn't work (anymore?) to have the same symbol
declared in multiple object files without initializing it. So instead
be careful to declare them all as extern and then separately declare
them in the c files.

It worked in CI, too, so I'm pretty confused, but oh well.

Also fix our attempts to keep it from doing 32-bit builds on OS X.