Make tuple structs identified/named based on runtime representation (#646)
Give them names based on the runtime types they contain.

This means that tuples whose component types have the same runtime
representation will also have the same runtime representation, so we
can do runtime subtyping on tuples.

This is also in preparation for separate compilation where it will
form the basis of standard naming of tuples between modules.  (And
will be combined with some preprocessor stuff to ensure that it gets
defined once).